Add files via upload
In this e-tivity, we were  asked to determine the components of a linear/non-linear programming problem, and then to implement in Python a program that will solve that problem by using the scipy.optimize library, how we determine the components of a linear/non-linear programming problem and also, how we implement linear/non-linear programs in Python. The program is based on a real-life scenario, which is invented by the author of module leader and thus, the numbers provided are fictive.

The real-life scenario
Mazda Motor Corporation produces three models of its CX series: Mazda CX-3, Mazda CX-5, and Mazda CX-9 (for simplicity, in this scenario we ignore the new models like CX-30, CX-50, etc.). Market projections indicate an expected monthly demand of at least: 3500 cars of the CX-3 model, 5000 cars of the CX-5 model, and 2100 cars of the CX-9 model. Due to insufficient production capacity, no more than 4000 cars can be made of model CX-3, no more than 7000 cars can be made of model CX-5, and no more than 2500 cars can be made of model CX-9. Due to existing contracts with car dealers worldwide, a total of at least 12000 cars must be produced each month. Here is the financial result calculated as production costs – selling costs per model:

CX-3: EUR 3400 profit
CX-5: EUR 2500 loss
CX-9: EUR 7200 profit

How many cars of each Mazda CX model should be produced monthly, so to maximize the total net profit?

The author of this assignment drives Mazda CX-5, 2021, and he loves it! He is convinced this car is undervalued. :)